+++ Tests/Reqnroll.SystemTests/Plugins/VerifyPluginTest.cs
@@ -13,7 +13,7 @@ public class VerifyPluginTest : PluginsTestBase
         _testRunConfiguration.UnitTestProvider = UnitTestProvider.xUnit;
         _projectsDriver.AddNuGetPackage("Reqnroll.Verify", _currentVersionDriver.ReqnrollNuGetVersion);
         // use the minimum version of xunit that is compatible with the Verify plugin
-        _solutionDriver.DefaultProject.UpdateNuGetPackage("xunit", "2.8.0");
+        _solutionDriver.DefaultProject.UpdateNuGetPackage("xunit", "2.8.1");
     }
 
     [TestMethod]